// src/app/models/admin.ts
export class test {
    product_id: number;
    price: number;
    quality:number
}

// export class users{
//     id: number;
//     name: string;
//     email: string;
//     password: string;
  
//   }

// export class category_product{
//     id: number;
//     name: string;
//     product_supplier_id:number;
//     warehouse_id: number;
//     status:number;

// }